refactor(reborn): hoist WebChatV2Event/Frame into ironclaw_product_workflow::webchat_schema
Behavior-preserving move so a future thin client (Reborn TUI) can depend on
the wire schema without depending on ironclaw_webui_v2. webui_v2::schema
re-exports; handlers.rs is unchanged. New snapshot test locks the wire
shape independent of the webui_v2 route surface.
